functions/archive_gcs/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.14.0
